Reduce the number of Native calls
Renamed HeimdalNtlm to Ntlm
Renamed HeimdalNtlmException to NtlmException
Naming convention in the shim layer is NetNativeNtlm_<Functionality>
