+++ src/Common/src/Interop/Linux/System.Net.Ntlm.Native/Interop.NtlmType3Message.cs
@@ -14,7 +14,7 @@ using Microsoft.Win32.SafeHandles;
 
 internal static partial class Interop
 {
-    internal static partial class HeimdalNtlm
+    internal static partial class Ntlm
     {
         /// <summary>
         /// Provides ntlm_type3 message
@@ -30,8 +30,8 @@ internal static partial class Interop
                 Debug.Assert(count >= 0 , " count must be a valid value");
                 Debug.Assert(type2Data.Length >= offset + count, " count and offset must match the given buffer");
 
-                int status = Interop.NetNtlmNative.HeimNtlmDecodeType2(type2Data, offset, count, out _type2Handle);
-                Interop.NetNtlmNative.HeimdalNtlmException.ThrowIfError(status);
+                int status = Interop.NetNtlmNative.NtlmDecodeType2(type2Data, offset, count, out _type2Handle);
+                Interop.NetNtlmNative.NtlmException.ThrowIfError(status);
             }
 
             public byte[] GetResponse(uint flags, string username, string password, string domain,
@@ -41,41 +41,21 @@ internal static partial class Interop
                 Debug.Assert(password != null, "password cannot be null");
                 Debug.Assert(domain != null,  "domain cannot be null");
 
-                // reference for NTLM response: https://msdn.microsoft.com/en-us/library/cc236700.aspx
-
                 sessionKey = null;
-                Interop.NetNtlmNative.NtlmBuffer key = default(Interop.NetNtlmNative.NtlmBuffer);
-                Interop.NetNtlmNative.NtlmBuffer lmResponse = default(Interop.NetNtlmNative.NtlmBuffer);
-                Interop.NetNtlmNative.NtlmBuffer ntResponse = default(Interop.NetNtlmNative.NtlmBuffer);
                 Interop.NetNtlmNative.NtlmBuffer sessionKeyBuffer = default(Interop.NetNtlmNative.NtlmBuffer);
                 Interop.NetNtlmNative.NtlmBuffer outputData = default(Interop.NetNtlmNative.NtlmBuffer);
 
                 try
                 {
-                    int status = Interop.NetNtlmNative.HeimNtlmNtKey(password, ref key);
-                    Interop.NetNtlmNative.HeimdalNtlmException.ThrowIfError(status);
-
-                    byte[] baseSessionKey = new byte[Interop.NetNtlmNative.MD5DigestLength];
-                    status = Interop.NetNtlmNative.HeimNtlmCalculateResponse(true, ref key, _type2Handle, username, domain,
-                             baseSessionKey, baseSessionKey.Length, ref lmResponse);
-                    Interop.NetNtlmNative.HeimdalNtlmException.ThrowIfError(status);
-
-                    status = Interop.NetNtlmNative.HeimNtlmCalculateResponse(false, ref key, _type2Handle, username, domain,
-                                                                           baseSessionKey, baseSessionKey.Length, ref ntResponse);
-                    Interop.NetNtlmNative.HeimdalNtlmException.ThrowIfError(status);
-
-                    status = Interop.NetNtlmNative.CreateType3Message(ref key, _type2Handle, username, domain, flags,
-                        ref lmResponse, ref ntResponse, baseSessionKey,baseSessionKey.Length, ref sessionKeyBuffer, ref outputData);
-                    Interop.NetNtlmNative.HeimdalNtlmException.ThrowIfError(status);
+                    int status = Interop.NetNtlmNative.CreateType3Message(password, _type2Handle, username, domain, flags,
+                        ref sessionKeyBuffer, ref outputData);
+                    Interop.NetNtlmNative.NtlmException.ThrowIfError(status);
 
                     sessionKey = sessionKeyBuffer.ToByteArray();
                     return outputData.ToByteArray();
                 }
                 finally
                 {
-                    key.Dispose();
-                    lmResponse.Dispose();
-                    ntResponse.Dispose();
                     sessionKeyBuffer.Dispose();
                     outputData.Dispose();
                 }
