+++ src/Native/System.Net.Ntlm.Native/pal_ntlmapi.h
@@ -32,46 +32,25 @@ extern "C" void NetNtlmNative_ReleaseNtlmBuffer(void* buffer, uint64_t length);
 /*
 Shims heim_ntlm_encode_type1 method.
 */
-extern "C" int32_t NetNtlmNative_HeimNtlmEncodeType1(uint32_t flags, struct PAL_NtlmBuffer* outBuffer);
+extern "C" int32_t NetNtlmNative_NtlmEncodeType1(uint32_t flags, struct PAL_NtlmBuffer* outBuffer);
 
 /*
 Shims heim_ntlm_decode_type2 method.
 */
-extern "C" int32_t NetNtlmNative_HeimNtlmDecodeType2(uint8_t* data, int32_t offset, int32_t count, ntlm_type2** type2);
+extern "C" int32_t NetNtlmNative_NtlmDecodeType2(uint8_t* data, int32_t offset, int32_t count, ntlm_type2** type2);
 
 /*
 Shims heim_ntlm_free_type2 method.
 */
-extern "C" void NetNtlmNative_HeimNtlmFreeType2(ntlm_type2* type2);
-
-/*
-Shims heim_ntlm_nt_key method.
-*/
-extern "C" int32_t NetNtlmNative_HeimNtlmNtKey(const char* password, struct PAL_NtlmBuffer* outBuffer);
-
-/*
-Shims heim_ntlm_calculate_lm2/_ntlm2 methods.
-*/
-extern "C" int32_t NetNtlmNative_HeimNtlmCalculateResponse(int32_t isLM,
-                                                           const struct PAL_NtlmBuffer* key,
-                                                           ntlm_type2* type2,
-                                                           char* username,
-                                                           char* target,
-                                                           uint8_t* baseSessionKey,
-                                                           int32_t baseSessionKeyLen,
-                                                           struct PAL_NtlmBuffer* data);
+extern "C" void NetNtlmNative_NtlmFreeType2(ntlm_type2* type2);
 
 /*
 Implements Type3 msg proccessing logic
 */
-extern "C" int32_t NetNtlmNative_CreateType3Message(struct PAL_NtlmBuffer* key,
+extern "C" int32_t NetNtlmNative_CreateType3Message(const char* password,
                                                     ntlm_type2* type2,
-                                                    const char* username,
-                                                    const char* domain,
+                                                    char* username,
+                                                    char* domain,
                                                     uint32_t flags,
-                                                    struct PAL_NtlmBuffer* lmResponse,
-                                                    struct PAL_NtlmBuffer* ntlmResponse,
-                                                    uint8_t* baseSessionKey,
-                                                    int32_t baseSessionKeyLen,
                                                     struct PAL_NtlmBuffer* outSessionKey,
-                                                    struct PAL_NtlmBuffer* outBufferHandle);
+                                                    struct PAL_NtlmBuffer* outBuffer);
