+++ src/System.Net.Security/src/Resources/Strings.resx
@@ -443,5 +443,7 @@
   </data>
   <data name="net_generic_heimntlm_operation_failed" xml:space="preserve">
     <value>Heimntlm operation failed with status: {0})</value>
+  <data name="net_generic_ntlm_operation_failed" xml:space="preserve">
+    <value>NTLM  operation failed with status: {0})</value>
   </data>
 </root>
